Fix missing QueryOptions delegations and maxIvaratorSources type mismatch
- Add missing delegation methods to QueryOptions for ivaratorFstHdfsBaseURIs,
  ivaratorPersistVerify, and ivaratorPersistVerifyCount
- Change maxIvaratorSources from long back to int in IvaratorConfig to match
  the ticket spec and avoid lossy (int) casts in SQC and QueryOptions

+++ warehouse/query-core/src/main/java/datawave/query/config/IvaratorConfig.java
@@ -44,7 +44,7 @@ public class IvaratorConfig implements Serializable {
     private int ivaratorNumRetries = 2;
     private boolean ivaratorPersistVerify = true;
     private int ivaratorPersistVerifyCount = 100;
-    private long maxIvaratorSources = 33;
+    private int maxIvaratorSources = 33;
     private long maxIvaratorSourceWait = 1000L * 60 * 30;
     private long maxIvaratorResults = -1;
 
@@ -175,11 +175,11 @@ public class IvaratorConfig implements Serializable {
         return this;
     }
 
-    public long getMaxIvaratorSources() {
+    public int getMaxIvaratorSources() {
         return maxIvaratorSources;
     }
 
-    public IvaratorConfig setMaxIvaratorSources(long maxIvaratorSources) {
+    public IvaratorConfig setMaxIvaratorSources(int maxIvaratorSources) {
         this.maxIvaratorSources = maxIvaratorSources;
         return this;
     }
